Align more closely with Gecko for clip-path reference box.
This changes how we calculate the reference box when the layout box is a
non-atomic inline.

It was very broken for LayoutNG; we just set the height to 0. For legacy
layout we used the width of the bounding box for all lines and the
height of the first fragment.

Instead, consistently use the size of the first fragment, in both NG and
legacy. Also get it right if writing-mode is vertical-rl.

Bug: 641907
Cq-Include-Trybots: luci.chromium.try:linux_layout_tests_layout_ng;luci.chromium.try:linux_layout_tests_slimming_paint_v2;master.tryserver.blink:linux_trusty_blink_rel
Change-Id: I652f197bc3d950fad1bf6067dc5f38c2df72a0a3